requirements.txt
django==1.9
djangorestframework
django-npm